fixed log bug and fixed avgpos calculation

+++ include/ERG_Mode.h
@@ -106,7 +106,7 @@ class PowerTable {
   void newEntry(PowerBuffer& powerBuffer);
 
   // enters data into power table 
-  void enterData(int i, int j, float pos); 
+  void enterData(int i, int j, int pos); 
 
   // returns incline for wattTarget. Null if not found.
   int32_t lookup(int watts, int cad);
